requirements.txt
torch==2.0.1
torchvision==0.15.2
pytorch_wavelets==0.3.3
onnx==1.14.1
onnxruntime==1.18.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.0
